chore: move utils to commmon
