+++ packages/components/textarea/textarea.tsx
@@ -14,7 +14,7 @@ import { isObject, merge, omit } from 'lodash-es';
 
 import calcTextareaHeight from './calcTextareaHeight';
 import { FormItemInjectionKey } from '../form/const';
-import setStyle from '../../common/js/utils/set-style';
+import setStyle from '../../common/js/utils/setStyle';
 import { getCharacterLength } from '../../common/js/utils/helper';
 
 // hooks
@@ -28,7 +28,7 @@ import useLengthLimit from '../input/useLengthLimit';
 import props from './props';
 import type { TextareaValue, TdTextareaProps } from './type';
 
-import { getValidAttrs } from '../../utils/helper';
+import { getValidAttrs } from '../../common/js/utils/helper';
 
 export default defineComponent({
   name: 'TTextarea',
